+++ views/menus/admMenu/contractManagerAdmMenu/contractManagerAdmMenu.c
@@ -48,9 +48,9 @@ void contractManagerAdmMenuChoose(){
         break;
         
         case 2:{
-            double contractId;
+            int contractId;
             printColorful("Informe o ID do contrato: ", 3);
-            scanf("%lf", &contractId);
+            scanf("%d", &contractId);
             
             if(!contractExistsById(contractId)){
                 printColorful("Contrato não encontrado.\n\n", 1);
@@ -103,11 +103,11 @@ void contractManagerAdmMenuChoose(){
         }
         
         case 5:{
-            double contractId;
+            int contractId;
             int contractStatus;
             
             printColorful("Informe o ID do contrato: ", 3);
-            scanf("%lf", &contractId);
+            scanf("%d", &contractId);
             
             if(!contractExistsById(contractId)){
                 printColorful("Contrato não encontrado.\n\n", 1);
@@ -132,9 +132,9 @@ void contractManagerAdmMenuChoose(){
         }
         
         case 6: {
-            double contractId;
+            int contractId;
             printColorful("Informe o ID do contrato: ", 3);
-            scanf("%lf", &contractId);
+            scanf("%d", &contractId);
             
             if(!contractExistsById(contractId)){
                 printColorful("Contrato não encontrado.\n\n", 1);
